backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
PyPDF2
python-docx
firebase-admin
